+++ test/test_with_xpress.jl
@@ -186,7 +186,7 @@ end
         chp_total_elec_prod = results["CHP"]["year_one_electric_production_series_kw"]
         chp_to_load = results["CHP"]["year_one_to_load_series_kw"]
         chp_export = results["CHP"]["year_one_to_grid_series_kw"]
-        cooling_elec_consumption = results["ExistingChiller"]["year_one_electric_consumption_series"]
+        cooling_elec_consumption = results["ExistingChiller"]["year_one_electric_consumption_series_kw"]
     
         # The values compared to the expected values
         @test sum([(chp_to_load[i] - tot_elec_load[i]*data["ElectricLoad"]["critical_load_fraction"]) for i in outage_start:outage_end]) ≈ 0.0 atol=0.001
@@ -464,7 +464,7 @@ end
     @test round(sum(r["ElectricStorage"]["year_one_soc_series_fraction"]), digits=2) / 8760 >= 0.72
 end
 
-@testset "Outage with Generator, outate simulator, BAU critical load outputs" begin
+@testset "Outage with Generator, outage simulator, BAU critical load outputs" begin
     m1 = Model(optimizer_with_attributes(Xpress.Optimizer, "OUTPUTLOG" => 0))
     m2 = Model(optimizer_with_attributes(Xpress.Optimizer, "OUTPUTLOG" => 0))
     p = REoptInputs("./scenarios/generator.json")
@@ -641,7 +641,7 @@ end
             p.s.dhw_load.loads_kw[ts] = 5
             p.s.space_heating_load.loads_kw[ts] = 5
             p.s.cooling_load.loads_kw_thermal[ts] = 10
-            p.s.existing_boiler.fuel_cost_series[ts] = 100
+            p.fuel_cost_per_kwh["ExistingBoiler"][ts] = 100
             for tier in 1:p.s.electric_tariff.n_energy_tiers
                 p.s.electric_tariff.energy_rates[ts, tier] = 100
             end
@@ -650,7 +650,7 @@ end
             p.s.dhw_load.loads_kw[ts] = 0
             p.s.space_heating_load.loads_kw[ts] = 0
             p.s.cooling_load.loads_kw_thermal[ts] = 0
-            p.s.existing_boiler.fuel_cost_series[ts] = 1
+            p.fuel_cost_per_kwh["ExistingBoiler"][ts] = 1
             for tier in 1:p.s.electric_tariff.n_energy_tiers
                 p.s.electric_tariff.energy_rates[ts, tier] = 100
             end
@@ -921,6 +921,27 @@ end
     cooling_electric_hybrid_expected = elec_hybrid .* (cooling_elec_frac_of_total_hospital * hospital_fraction  .+ 
                                             cooling_elec_frac_of_total_hotel * hotel_fraction)
     @test round(sum(cooling_electric_hybrid_expected .- cooling_elec_hybrid), digits=1) ≈ 0.0 atol=0.1
+end
+
+@testset "Boiler (new) test" begin
+    input_data = JSON.parsefile("scenarios/boiler_new_inputs.json")
+    input_data["SpaceHeatingLoad"]["annual_mmbtu"] = 0.5 * 8760
+    input_data["DomesticHotWaterLoad"]["annual_mmbtu"] = 0.5 * 8760
+    s = Scenario(input_data)
+    inputs = REoptInputs(s)
+    m1 = Model(optimizer_with_attributes(Xpress.Optimizer, "OUTPUTLOG" => 0))
+    m2 = Model(optimizer_with_attributes(Xpress.Optimizer, "OUTPUTLOG" => 0))
+    results = run_reopt([m1,m2], inputs)
+    
+    # BAU boiler loads
+    load_thermal_mmbtu_bau = sum(s.space_heating_load.loads_kw + s.dhw_load.loads_kw) / REopt.KWH_PER_MMBTU
+    existing_boiler_mmbtu = sum(results["ExistingBoiler"]["year_one_thermal_production_series_mmbtu_per_hour"])
+    boiler_thermal_mmbtu = sum(results["Boiler"]["year_one_thermal_production_series_mmbtu_per_hour"])
+    
+    # Used monthly fuel cost for ExistingBoiler and Boiler, where ExistingBoiler has lower fuel cost only
+    # in February (28 days), so expect ExistingBoiler to serve the flat/constant load 28 days of the year
+    @test existing_boiler_mmbtu ≈ load_thermal_mmbtu_bau * 28 / 365 atol=0.00001
+    @test boiler_thermal_mmbtu ≈ load_thermal_mmbtu_bau - existing_boiler_mmbtu atol=0.00001
 end
 
 @testset "OffGrid" begin
@@ -1254,57 +1275,168 @@ end
     end
 end
 
+@testset "Back pressure steam turbine" begin
+    """
+    Validation to ensure that:
+        1) ExistingBoiler provides the thermal energy (steam) to a backpressure SteamTurbine for CHP application
+        2) SteamTurbine serves the heating load with the condensing steam
 
-## equivalent REopt API Post for test 2:
-#   NOTE have to hack in API levelization_factor to get LCC within 5e-5 (Mosel tol)
-# {"Scenario": {
-#     "Site": {
-#         "longitude": -118.1164613,
-#         "latitude": 34.5794343,
-#         "roof_squarefeet": 5000.0,
-#         "land_acres": 1.0,
-#     "PV": {
-#         "macrs_bonus_fraction": 0.4,
-#         "installed_cost_per_kw": 2000.0,
-#         "tilt": 34.579,
-#         "degradation_fraction": 0.005,
-#         "macrs_option_years": 5,
-#         "federal_itc_fraction": 0.3,
-#         "module_type": 0,
-#         "array_type": 1,
-#         "om_cost_per_kw": 16.0,
-#         "macrs_itc_reduction": 0.5,
-#         "azimuth": 180.0,
-#         "federal_rebate_per_kw": 350.0,
-#         "dc_ac_ratio": 1.1
-#     },
-#     "LoadProfile": {
-#         "doe_reference_name": "RetailStore",
-#         "annual_kwh": 10000000.0,
-#         "city": "LosAngeles"
-#     },
-#     "ElectricStorage": {
-#         "total_rebate_per_kw": 100.0,
-#         "macrs_option_years": 5,
-#         "can_grid_charge": true,
-#         "macrs_bonus_fraction": 0.4,
-#         "macrs_itc_reduction": 0.5,
-#         "total_itc_fraction": 0,
-#         "installed_cost_per_kw": 1000.0,
-#         "installed_cost_per_kwh": 500.0,
-#         "replace_cost_per_kw": 460.0,
-#         "replace_cost_per_kwh": 230.0
-#     },
-#     "ElectricTariff": {
-#         "urdb_label": "5ed6c1a15457a3367add15ae"
-#     },
-#     "Financial": {
-#         "escalation_rate_fraction": 0.026,
-#         "offtaker_discount_rate_fraction": 0.081,
-#         "owner_discount_rate_fraction": 0.081,
-#         "analysis_years": 20,
-#         "offtaker_tax_rate_fraction": 0.4,
-#         "owner_tax_rate_fraction": 0.4,
-#         "om_cost_escalation_rate_fraction": 0.025
-#     }
-# }}}
+    """
+    # Setup inputs, make heating load large to entice SteamTurbine
+    input_data = JSON.parsefile("scenarios/backpressure_steamturbine_inputs.json")
+    latitude = input_data["Site"]["latitude"]
+    longitude = input_data["Site"]["longitude"]
+    building = "Hospital"
+    elec_load_multiplier = 5.0
+    heat_load_multiplier = 100.0
+    input_data["ElectricLoad"]["doe_reference_name"] = building
+    input_data["SpaceHeatingLoad"]["doe_reference_name"] = building
+    input_data["DomesticHotWaterLoad"]["doe_reference_name"] = building
+    elec_load = REopt.ElectricLoad(latitude=latitude, longitude=longitude, doe_reference_name=building)
+    input_data["ElectricLoad"]["annual_kwh"] = elec_load_multiplier * sum(elec_load.loads_kw)
+    space_load = REopt.SpaceHeatingLoad(latitude=latitude, longitude=longitude, doe_reference_name=building)
+    input_data["SpaceHeatingLoad"]["annual_mmbtu"] = heat_load_multiplier * space_load.annual_mmbtu / input_data["ExistingBoiler"]["efficiency"]
+    dhw_load = REopt.DomesticHotWaterLoad(latitude=latitude, longitude=longitude, doe_reference_name=building)
+    input_data["DomesticHotWaterLoad"]["annual_mmbtu"] = heat_load_multiplier * dhw_load.annual_mmbtu / input_data["ExistingBoiler"]["efficiency"]
+    s = Scenario(input_data)
+    inputs = REoptInputs(s)
+    m1 = Model(optimizer_with_attributes(Xpress.Optimizer, "OUTPUTLOG" => 0))
+    m2 = Model(optimizer_with_attributes(Xpress.Optimizer, "OUTPUTLOG" => 0))
+    results = run_reopt([m1,m2], inputs)
+
+    # The expected values below were directly copied from the REopt_API V2 expected values
+    @test results["Financial"]["lcc"] ≈ 189359280.0 rtol=0.001
+    @test results["Financial"]["npv"] ≈ 8085233.0 rtol=0.01
+    @test results["SteamTurbine"]["size_kw"] ≈ 2616.418 atol=1.0
+    @test results["SteamTurbine"]["year_one_thermal_consumption_mmbtu"] ≈ 1000557.6 rtol=0.001
+    @test results["SteamTurbine"]["year_one_electric_energy_produced_kwh"] ≈ 18970374.6 rtol=0.001
+    @test results["SteamTurbine"]["year_one_thermal_energy_produced_mmbtu"] ≈ 924045.1 rtol=0.001
+
+    # BAU boiler loads
+    load_boiler_fuel = (s.space_heating_load.loads_kw + s.dhw_load.loads_kw) ./ REopt.KWH_PER_MMBTU ./ s.existing_boiler.efficiency
+    load_boiler_thermal = load_boiler_fuel * s.existing_boiler.efficiency
+
+    # ExistingBoiler and SteamTurbine production
+    boiler_to_load = results["ExistingBoiler"]["year_one_thermal_to_load_series_mmbtu_per_hour"]
+    boiler_to_st = results["ExistingBoiler"]["year_one_thermal_to_steamturbine_series_mmbtu_per_hour"]
+    boiler_total = boiler_to_load + boiler_to_st
+    st_to_load = results["SteamTurbine"]["year_one_thermal_to_load_series_mmbtu_per_hour"]
+
+    # Fuel/thermal **consumption**
+    boiler_fuel = results["ExistingBoiler"]["year_one_fuel_consumption_series_mmbtu_per_hour"]
+    steamturbine_thermal_in = results["SteamTurbine"]["year_one_thermal_consumption_series_mmbtu_per_hour"]
+
+    # Check that all thermal supply to load meets the BAU load
+    thermal_to_load = sum(boiler_to_load) + sum(st_to_load)
+    @test thermal_to_load ≈ sum(load_boiler_thermal) atol=1.0
+
+    # Check the net electric efficiency of Boiler->SteamTurbine (electric out/fuel in) with the expected value from the Fact Sheet 
+    steamturbine_electric = results["SteamTurbine"]["year_one_electric_production_series_kw"] 
+    net_electric_efficiency = sum(steamturbine_electric) / (sum(boiler_fuel) * REopt.KWH_PER_MMBTU)
+    @test net_electric_efficiency ≈ 0.052 atol=0.005
+
+    # Check that the max production of the boiler is still less than peak heating load times thermal factor
+    factor = input_data["ExistingBoiler"]["max_thermal_factor_on_peak_load"]
+    boiler_capacity = maximum(load_boiler_thermal) * factor
+    @test maximum(boiler_total) <= boiler_capacity
+end
+
+@testset "All heating supply/demand/storage energy balance" begin
+    """
+    Validation to ensure that:
+        1) Heat balance is correct with SteamTurbine (backpressure), CHP, HotTES, and AbsorptionChiller included
+        2) The sum of a all thermal from techs supplying SteamTurbine is equal to SteamTurbine thermal consumption
+        3) Techs are not supplying SteamTurbine with thermal if can_supply_steam_turbine = False
+    
+    :return:
+    """
+    
+    # Start with steam turbine inputs, but adding a bunch below
+    input_data = JSON.parsefile("scenarios/backpressure_steamturbine_inputs.json")
+    input_data["ElectricLoad"]["doe_reference_name"] = "Hospital"
+    # Add SpaceHeatingLoad building for heating loads, ignore DomesticHotWaterLoad for simplicity of energy balance checks
+    input_data["SpaceHeatingLoad"]["doe_reference_name"] = "Hospital"
+    delete!(input_data, "DomesticHotWaterLoad")
+    
+    # Fix size of SteamTurbine, even if smaller than practical, because we're just looking at energy balances
+    input_data["SteamTurbine"]["min_kw"] = 30.0
+    input_data["SteamTurbine"]["max_kw"] = 30.0
+    
+    # Add CHP 
+    input_data["CHP"] = Dict{Any, Any}([
+                        ("prime_mover", "recip_engine"),
+                        ("size_class", 2),
+                        ("min_kw", 250.0),
+                        ("min_allowable_kw", 0.0),
+                        ("max_kw", 250.0),
+                        ("can_supply_steam_turbine", false),
+                        ("fuel_cost_per_mmbtu", 8.0),
+                        ("cooling_thermal_factor", 1.0)
+                        ])
+    
+    input_data["Financial"]["chp_fuel_cost_escalation_rate_fraction"] = 0.034
+    
+    # Add CoolingLoad and AbsorptionChiller so we can test the energy balance on AbsorptionChiller too (thermal consumption)
+    input_data["CoolingLoad"] = Dict{Any, Any}("doe_reference_name" => "Hospital")
+    input_data["AbsorptionChiller"] = Dict{Any, Any}([
+                                        ("min_ton", 600.0),
+                                        ("max_ton", 600.0),
+                                        ("cop_thermal", 0.7),
+                                        ("installed_cost_per_ton", 500.0),
+                                        ("om_cost_per_ton", 0.5)
+                                        ])
+    
+    # Add Hot TES
+    input_data["HotThermalStorage"] = Dict{Any, Any}([
+                            ("min_gal", 50000.0),
+                            ("max_gal", 50000.0)
+                            ])
+    
+    s = Scenario(input_data)
+    inputs = REoptInputs(s)
+    m = Model(optimizer_with_attributes(Xpress.Optimizer, "MIPRELSTOP" => 0.01, "OUTPUTLOG" => 0))
+    results = run_reopt(m, inputs)
+    
+    thermal_techs = ["ExistingBoiler", "CHP", "SteamTurbine"]
+    thermal_loads = ["load", "tes", "steamturbine", "waste"]  # We don't track AbsorptionChiller thermal consumption by tech
+    tech_to_thermal_load = Dict{Any, Any}()
+    for tech in thermal_techs
+        tech_to_thermal_load[tech] = Dict{Any, Any}()
+        for load in thermal_loads
+            if (tech == "SteamTurbine" && load == "steamturbine") || (load == "waste" && tech != "CHP")
+                tech_to_thermal_load[tech][load] = [0.0] * 8760
+            else
+                tech_to_thermal_load[tech][load] = results[tech]["year_one_thermal_to_"*load*"_series_mmbtu_per_hour"]
+            end
+        end
+    end
+    # Hot TES is the other thermal supply
+    hottes_to_load = results["HotThermalStorage"]["year_one_to_load_series_mmbtu_per_hour"]
+    
+    # BAU boiler loads
+    load_boiler_fuel = s.space_heating_load.loads_kw / input_data["ExistingBoiler"]["efficiency"] ./ REopt.KWH_PER_MMBTU
+    load_boiler_thermal = load_boiler_fuel .* REopt.EXISTING_BOILER_EFFICIENCY
+    
+    # Fuel/thermal **consumption**
+    boiler_fuel = results["ExistingBoiler"]["year_one_fuel_consumption_series_mmbtu_per_hour"]
+    chp_fuel_total = results["CHP"]["year_one_fuel_used_mmbtu"]
+    steamturbine_thermal_in = results["SteamTurbine"]["year_one_thermal_consumption_series_mmbtu_per_hour"]
+    absorptionchiller_thermal_in = results["AbsorptionChiller"]["year_one_thermal_consumption_series_mmbtu_per_hour"]
+    
+    # Check that all thermal supply to load meets the BAU load plus AbsorptionChiller load which is not explicitly tracked
+    alltechs_thermal_to_load_total = sum([sum(tech_to_thermal_load[tech]["load"]) for tech in thermal_techs]) + sum(hottes_to_load)
+    thermal_load_total = sum(load_boiler_thermal) + sum(absorptionchiller_thermal_in)
+    @test alltechs_thermal_to_load_total ≈ thermal_load_total atol=0.01
+    
+    # Check that all thermal to steam turbine is equal to steam turbine thermal consumption
+    alltechs_thermal_to_steamturbine_total = sum([sum(tech_to_thermal_load[tech]["steamturbine"]) for tech in ["ExistingBoiler", "CHP"]])
+    @test alltechs_thermal_to_steamturbine_total ≈ sum(steamturbine_thermal_in) atol=3
+    
+    # Check that "thermal_to_steamturbine" is zero for each tech which has input of can_supply_steam_turbine as False
+    for tech in ["ExistingBoiler", "CHP"]
+        if !(tech in inputs.techs.can_supply_steam_turbine)
+            @test sum(tech_to_thermal_load[tech]["steamturbine"]) == 0.0
+        end
+    end
+end
+